fix  : 반응형 UI 수정 및 모바일 화면 수정
